Refactor Registrar Actions API
Extract shared components between v0 and v1
